remove unused change providers
