+++ corehq/pillows/ledger.py
@@ -3,14 +3,13 @@ from corehq.apps.change_feed.consumer.feed import KafkaChangeFeed
 from corehq.apps.locations.models import SQLLocation
 from corehq.elastic import get_es_new
 from corehq.form_processor.backends.sql.dbaccessors import LedgerReindexAccessor
-from corehq.form_processor.change_providers import (
-    _ledger_v1_to_change
-)
+from corehq.form_processor.change_publishers import change_meta_from_ledger_v1
 from corehq.form_processor.utils.general import should_use_sql_backend
 from corehq.pillows.mappings.ledger_mapping import LEDGER_INDEX_INFO
 from corehq.util.doc_processor.sql import SqlDocumentProvider
 from corehq.util.quickcache import quickcache
 from pillowtop.checkpoints.manager import PillowCheckpoint, PillowCheckpointEventHandler
+from pillowtop.feed.interface import Change
 from pillowtop.pillow.interface import ConstructedPillow
 from pillowtop.processors.elastic import ElasticProcessor
 from pillowtop.reindexer.change_providers.django_model import DjangoModelChangeProvider
@@ -94,4 +93,15 @@ def get_ledger_v1_reindexer():
         change_provider=DjangoModelChangeProvider(StockState, _ledger_v1_to_change),
         elasticsearch=get_es_new(),
         index_info=LEDGER_INDEX_INFO,
+    )
+
+
+def _ledger_v1_to_change(stock_state):
+    return Change(
+        id=stock_state.pk,
+        sequence_id=None,
+        document=stock_state.to_json(),
+        deleted=False,
+        metadata=change_meta_from_ledger_v1(stock_state),
+        document_store=None,
     )
